// examples/typescript/gen/valorem/trade/v1/auth_connect.ts
// @generated by protoc-gen-connect-es v0.9.1 with parameter "target=ts,import_extension=none"
// @generated from file auth.proto (package quay, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { Empty, NonceText, VerifyText } from "./auth_pb";
import { MethodKind } from "@bufbuild/protobuf";
import { H160 } from "./types_pb";

/**
 * @generated from service quay.Auth
 */
export const Auth = {
  typeName: "quay.Auth",
  methods: {
    /**
     * * Returns an EIP-4361 nonce for session and invalidates existing session
     *
     * @generated from rpc quay.Auth.Nonce
     */
    nonce: {
      name: "Nonce",
      I: Empty,
      O: NonceText,
      kind: MethodKind.Unary,
    },
    /**
     * * Verifies the SignedMessage is valid and returns the verified address
     *
     * @generated from rpc quay.Auth.Verify
     */
    verify: {
      name: "Verify",
      I: VerifyText,
      O: H160,
      kind: MethodKind.Unary,
    },
    /**
     * * Used to check if a given connection is authenticated, returns the address which is authenticated for a nonce cookie
     *
     * @generated from rpc quay.Auth.Authenticate
     */
    authenticate: {
      name: "Authenticate",
      I: Empty,
      O: H160,
      kind: MethodKind.Unary,
    },
  }
} as const;