// examples/typescript/gen/valorem/trade/v1/rfq_connect.ts
// @generated by protoc-gen-connect-es v0.9.1 with parameter "target=ts,import_extension=none"
// @generated from file rfq.proto (package quay, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import { QuoteRequest, QuoteResponse } from "./rfq_pb";
import { MethodKind } from "@bufbuild/protobuf";

/**
 * @generated from service quay.RFQ
 */
export const RFQ = {
  typeName: "quay.RFQ",
  methods: {
    /**
     * * Request quotes from makers via a single QuoteRequest message and receive a stream of QuoteResponse messages for use by gRPC-web clients. 
     *
     * @generated from rpc quay.RFQ.WebTaker
     */
    webTaker: {
      name: "WebTaker",
      I: QuoteRequest,
      O: QuoteResponse,
      kind: MethodKind.ServerStreaming,
    },
    /**
     * * Request quotes from makers via a stream of QuoteRequest messages and receive a stream of QuoteResponse messages. 
     *
     * @generated from rpc quay.RFQ.Taker
     */
    taker: {
      name: "Taker",
      I: QuoteRequest,
      O: QuoteResponse,
      kind: MethodKind.BiDiStreaming,
    },
    /**
     * * Send quotes to takers via a stream of QuoteResponse messages and receive a stream of QuoteRequest messages. 
     *
     * @generated from rpc quay.RFQ.Maker
     */
    maker: {
      name: "Maker",
      I: QuoteResponse,
      O: QuoteRequest,
      kind: MethodKind.BiDiStreaming,
    },
  }
} as const;