// examples/typescript/gen/valorem/trade/v1/types_pb.ts
export class EthSignature extends Message<EthSignature> {
  /**
   * @generated from field: bytes r = 1;
   */
  r = new Uint8Array(0);

  /**
   * @generated from field: bytes s = 2;
   */
  s = new Uint8Array(0);

  /**
   * @generated from field: bytes v = 3;
   */
  v = new Uint8Array(0);

  constructor(data?: PartialMessage<EthSignature>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "quay.EthSignature";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "r", kind: "scalar", T: 12 /* ScalarType.BYTES */ },
    { no: 2, name: "s", kind: "scalar", T: 12 /* ScalarType.BYTES */ },
    { no: 3, name: "v", kind: "scalar", T: 12 /* ScalarType.BYTES */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): EthSignature {
    return new EthSignature().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): EthSignature {
    return new EthSignature().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): EthSignature {
    return new EthSignature().fromJsonString(jsonString, options);
  }

  static equals(a: EthSignature | PlainMessage<EthSignature> | undefined, b: EthSignature | PlainMessage<EthSignature> | undefined): boolean {
    return proto3.util.equals(EthSignature, a, b);
  }
}