// examples/typescript/gen/valorem/trade/v1/seaport_pb.ts
export enum ItemType {
  /**
   * @generated from enum value: NATIVE = 0;
   */
  NATIVE = 0,

  /**
   * @generated from enum value: ERC20 = 1;
   */
  ERC20 = 1,

  /**
   * @generated from enum value: ERC721 = 2;
   */
  ERC721 = 2,

  /**
   * @generated from enum value: ERC1155 = 3;
   */
  ERC1155 = 3,

  /**
   * @generated from enum value: ERC721_WITH_CRITERIA = 4;
   */
  ERC721_WITH_CRITERIA = 4,

  /**
   * @generated from enum value: ERC1155_WITH_CRITERIA = 5;
   */
  ERC1155_WITH_CRITERIA = 5,
}
// ...
export enum OrderType {
  /**
   * @generated from enum value: FULL_OPEN = 0;
   */
  FULL_OPEN = 0,

  /**
   * @generated from enum value: PARTIAL_OPEN = 1;
   */
  PARTIAL_OPEN = 1,

  /**
   * @generated from enum value: FULL_RESTRICTED = 2;
   */
  FULL_RESTRICTED = 2,

  /**
   * @generated from enum value: PARTIAL_RESTRICTED = 3;
   */
  PARTIAL_RESTRICTED = 3,
}
// Retrieve enum metadata with: proto3.getEnumType(OrderType)
proto3.util.setEnumType(OrderType, "quay.OrderType", [
  { no: 0, name: "FULL_OPEN" },
  { no: 1, name: "PARTIAL_OPEN" },
  { no: 2, name: "FULL_RESTRICTED" },
  { no: 3, name: "PARTIAL_RESTRICTED" },
]);

/**
 * *
 * This is and item required in exchange for an offer.
 *
 * @generated from message quay.ConsiderationItem
 */
export class ConsiderationItem extends Message<ConsiderationItem> {
  /**
   * @generated from field: quay.ItemType item_type = 1;
   */
  itemType = ItemType.NATIVE;

  /**
   * @generated from field: quay.H160 token = 2;
   */
  token?: H160;

  /**
   * uint256
   *
   * @generated from field: quay.H256 identifier_or_criteria = 3;
   */
  identifierOrCriteria?: H256;

  /**
   * uint256
   *
   * @generated from field: quay.H256 start_amount = 4;
   */
  startAmount?: H256;

  /**
   * uint256
   *
   * @generated from field: quay.H256 end_amount = 5;
   */
  endAmount?: H256;

  /**
   * @generated from field: quay.H160 recipient = 6;
   */
  recipient?: H160;

  constructor(data?: PartialMessage<ConsiderationItem>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "quay.ConsiderationItem";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "item_type", kind: "enum", T: proto3.getEnumType(ItemType) },
    { no: 2, name: "token", kind: "message", T: H160 },
    { no: 3, name: "identifier_or_criteria", kind: "message", T: H256 },
    { no: 4, name: "start_amount", kind: "message", T: H256 },
    { no: 5, name: "end_amount", kind: "message", T: H256 },
    { no: 6, name: "recipient", kind: "message", T: H160 },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ConsiderationItem {
    return new ConsiderationItem().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ConsiderationItem {
    return new ConsiderationItem().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ConsiderationItem {
    return new ConsiderationItem().fromJsonString(jsonString, options);
  }

  static equals(a: ConsiderationItem | PlainMessage<ConsiderationItem> | undefined, b: ConsiderationItem | PlainMessage<ConsiderationItem> | undefined): boolean {
    return proto3.util.equals(ConsiderationItem, a, b);
  }
}

/**
 * *
 * This is an item offered in exchange for consideration.
 *
 * @generated from message quay.OfferItem
 */
export class OfferItem extends Message<OfferItem> {
  /**
   * @generated from field: quay.ItemType item_type = 1;
   */
  itemType = ItemType.NATIVE;

  /**
   * * The token designates the account of the item's token contract
   * (with the null address used for Ether or other native tokens).
   *
   * @generated from field: quay.H160 token = 2;
   */
  token?: H160;

  /**
   * * The identifier_or_criteria represents either the ERC721 or ERC1155
   * token identifier or, in the case of a criteria-based item type, a
   * merkle root composed of the valid set of token identifiers for
   * the item. This value will be ignored for Ether and ERC20 item types,
   * and can optionally be zero for criteria-based item types to allow
   * for any identifier.
   *
   * @generated from field: quay.H256 identifier_or_criteria = 3;
   */
  identifierOrCriteria?: H256;

  /**
   * * The start_amount represents the amount of the item in question that
   * will be required should the order be fulfilled at the moment the
   * order becomes active.
   *
   * @generated from field: quay.H256 start_amount = 4;
   */
  startAmount?: H256;

  /**
   * * The end_amount represents the amount of the item in question that
   * will be required should the order be fulfilled at the moment the
   * order expires. If this value differs from the item's start_amount,
   * the realized amount is calculated linearly based on the time elapsed
   * since the order became active.
   *
   * uint256
   *
   * @generated from field: quay.H256 end_amount = 5;
   */
  endAmount?: H256;

  constructor(data?: PartialMessage<OfferItem>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "quay.OfferItem";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "item_type", kind: "enum", T: proto3.getEnumType(ItemType) },
    { no: 2, name: "token", kind: "message", T: H160 },
    { no: 3, name: "identifier_or_criteria", kind: "message", T: H256 },
    { no: 4, name: "start_amount", kind: "message", T: H256 },
    { no: 5, name: "end_amount", kind: "message", T: H256 },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): OfferItem {
    return new OfferItem().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): OfferItem {
    return new OfferItem().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): OfferItem {
    return new OfferItem().fromJsonString(jsonString, options);
  }

  static equals(a: OfferItem | PlainMessage<OfferItem> | undefined, b: OfferItem | PlainMessage<OfferItem> | undefined): boolean {
    return proto3.util.equals(OfferItem, a, b);
  }
}
// ...
export class SignedOrder extends Message<SignedOrder> {
  /**
   * @generated from field: quay.Order parameters = 1;
   */
  parameters?: Order;

  /**
   * @generated from field: quay.EthSignature signature = 2;
   */
  signature?: EthSignature;

  constructor(data?: PartialMessage<SignedOrder>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "quay.SignedOrder";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "parameters", kind: "message", T: Order },
    { no: 2, name: "signature", kind: "message", T: EthSignature },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): SignedOrder {
    return new SignedOrder().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): SignedOrder {
    return new SignedOrder().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): SignedOrder {
    return new SignedOrder().fromJsonString(jsonString, options);
  }

  static equals(a: SignedOrder | PlainMessage<SignedOrder> | undefined, b: SignedOrder | PlainMessage<SignedOrder> | undefined): boolean {
    return proto3.util.equals(SignedOrder, a, b);
  }
}